requirements.txt
streamlit
opencv-python
pytesseract
easyocr
pandas
numpy
openai
Pillow